% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis.R
\name{Basis}
\alias{Basis}
\title{Basis class}
\description{
An `R6` class for a basis function in the style of `hal9001`.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orders}}{A numeric vector for the orders of the basis functions.}

\item{\code{cols}}{numeric vector for the column indices of the basis functions.}

\item{\code{cutoffs}}{A numeric vector for the cutoffs of the basis functions.
Initialize the Basis object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Basis-new}{\code{Basis$new()}}
\item \href{#method-Basis-clone}{\code{Basis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Basis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Basis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$new(orders, cols, cutoffs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orders}}{A numeric vector for the orders of the basis functions.}

\item{\code{cols}}{A numeric vector for the column indices of the basis functions.}

\item{\code{cutoffs}}{A numeric vector for the cutoffs of the basis functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Basis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Basis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
