% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R
\name{RV}
\alias{RV}
\title{RV (random variable) class}
\description{
An `R6` class for randomly generating a random variable.
Currently only normal, uniform, and binomial distributions are supported.
}
\examples{
set.seed(123)
normal_rv <- RV$new(dist_name = "normal", params = list(mean = 0, sd = 1))
normal_rv$generate_samples(5)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dist_name}}{Name of the distribution.}

\item{\code{params}}{Parameters for the distribution.}

\item{\code{vals}}{Realizations of the RV.}

\item{\code{range}}{Range of the RV.}

\item{\code{cont}}{Whether the RV is continuous.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RV-new}{\code{RV$new()}}
\item \href{#method-RV-generate_samples}{\code{RV$generate_samples()}}
\item \href{#method-RV-get_knots}{\code{RV$get_knots()}}
\item \href{#method-RV-clone}{\code{RV$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RV-new"></a>}}
\if{latex}{\out{\hypertarget{method-RV-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the RV class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RV$new(dist_name, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dist_name}}{A string for the name of the distribution.}

\item{\code{params}}{A list of parameters for the distribution.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RV-generate_samples"></a>}}
\if{latex}{\out{\hypertarget{method-RV-generate_samples}{}}}
\subsection{Method \code{generate_samples()}}{
Generate samples from the RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RV$generate_samples(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{An integer for the number of samples to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RV-get_knots"></a>}}
\if{latex}{\out{\hypertarget{method-RV-get_knots}{}}}
\subsection{Method \code{get_knots()}}{
Randomly draw k knots from the range of the RV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RV$get_knots(k)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{An integer for the number of knots to draw.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
